Fixing some windows errors.
Made some implict casts explicit so that MSVC won't issue warnings about
possible loss of information, and changed an instance of uint to
uint32_t. I didn't even know uint was a valid thing, but either way MSVC
doesn't like it.

+++ src/splat_game.cpp
@@ -275,8 +275,8 @@ bool SplatGame::InitializeGameState() {
   // Add a touch screen controller into the controller list, so that touch
   // inputs are processed correctly and assigned a character:
   TouchscreenController * ts_controller = new TouchscreenController();
-  vec2 window_size = vec2(renderer_.window_size().x(),
-                          renderer_.window_size().y());
+  vec2 window_size = vec2(static_cast<float>(renderer_.window_size().x()),
+                          static_cast<float>(renderer_.window_size().y()));
   ts_controller->Initialize(&input_, window_size, &config);
   AddController(ts_controller);
